Legg template i egen fil, importer dotenv
