+++ package.json
@@ -45,6 +45,7 @@
         "babel-preset-react-app": "^9.0.2",
         "classnames": "^2.2.6",
         "css-loader": "^3.2.0",
+        "dotenv": "^8.2.0",
         "downshift": "^3.2.10",
         "es6-promise": "^4.2.8",
         "express": "^4.17.1",
