+++ webpack.server.config.js
@@ -15,7 +15,7 @@ const browserConfig = {
         }),
     ],
     entry: {
-        server: path.resolve(__dirname, './src/server/server.tsx'),
+        server: path.resolve(__dirname, './src/server/server.ts'),
     },
     output: {
         path: path.resolve(__dirname, 'build/'),
